+++ backend/src/main/java/com/woowacourse/moamoa/auth/service/oauthclient/response/OAuthAccessTokenResponse.java
@@ -1,4 +1,4 @@
-package com.woowacourse.moamoa.auth.service.response;
+package com.woowacourse.moamoa.auth.service.oauthclient.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
